refactor: rremove filesystem.rs, move parse_github_url to utils.rs, and simplify main.rs
