+++ src/analyzers/mod.rs
@@ -1,0 +1,5 @@
+pub mod filesystem;
+pub mod repo;
+pub mod security;
+pub mod type_detector;
+pub mod code_metrics;
